Refactors code, renames container to manager
